Fixed navigation link color bug #517
Fixed bug by forcing active={false} in the Nav.Link properties. This forces the program to reset all links's active status every url change. This means that only the active url will have a active link status.

The program will still highlight in red what page the user is on but will revert all links back to black if the link to the main page or the user is on the main page.

I also did slight reformatting of the Navbar.